Created file for storing words
`words.py` will be used as a "database" of sorts to store words based on
themes. This can be expanded to include newer words for the same theme
or newer themes.

At present, the format is a simple list and the filetype is `.py`. This
can be modified to a json.