Hangman Algorithm
The core game logic whereby user gets 9 attempts.
As long as the word is not guessed, each guess is checked for validity.
If the guess is valid, the occurances of the guessed letter and their
corresponding indices are obtained. The dashes are replaced with the
correctly guessed letters and the game continues.

Once all dashes used up and the user still has lives left, the game is
considered won.

If user finishes up all attemps with blanks left, the user loses the
game.