Changes to Code Structure for Creating Executable
`hangman.py` was restructured to have a `__main__.py` file that now has
a `main()` method. Outside of the `hangman/`dir, a `cli.py` was created
that allows for `hangman/__main__.py` to be an entrypoint for the app.

An executable was created based on this new struture using PyInstaller.

To now run the code from the shell please call cli.py