Decouple guess and word in Game.start()
The start method in Game is coupled with two entities, the guess made
by the player and the word of the game. The main logic for the game
requires us to make decisions based on checked performed on both these
entities. The current implementation resulted in a tightly coupled code,
with start() performing too many tasks and making it difficult to
unittest single bits of functionality.

This commit proposes a fix in the form of introducing a Guess object
which encapsulates the validations we need to perform into methods that
are easily testable.

Additionally, some refactoring is done to Game.start() by removing the
guessed variable and pulling win and lost states into their own methods.