Testing Read Replay and Refactoring Read Replay Check
Previously logic to test for user input for replay was in the
`__main__.py`. This was moved into `interface.py` to maintain
consistency - all inputs are checked in the `interface.py` Class.
Finally, `test_read_replay` was created to test user input for reply to
be upper case and of length 1 no matter what.