[feature]: integration tests for game.start
A use case/scenario based approach is taken to perform tests on the
game.start method. The Game class functions as a controller with the
start method containing the main control logic, thus integration tests
are performed. The unittest.mock module is used to mock out the methods
that print to stdout.